stats_thread: switch from timer+signal to clock_nanosleep
Using timers+signals together with threads presents some tricky issues.
In this case, it caused problems with interupted system calls in libnl.
Since there's a dedicated stats thread, it made sense to make it synchronous.

This change makes libnl happy, but opens up multi-threading problems in
the fossa library. That problem will be fixed in a future patch, by splitting
the control and data connections.