bugfix: protect the fossa connection object with a mutex.
The stats_thread and jittertrap threads access the fossa objects
concurrently, but the fossa library is not thread-safe. This change adds
a massive big mutex around the event handler and stats_thread callbacks
to serialise the access. Performance seems adequate (it doesn't affect the
sampling accuracy) and buys some time while a better multi-process
architecture is being worked on.