optimisation: only recalculate stats for the visible series.
Profiling showed that a substantial amount of CPU time is consumed by
recalculating the basic min/max/median stats and histogram for each new
data point and it gets worse when new calculations like FFT are added.

This change removes the auxiliary stats calculations for the 7/8 data
series that are not being displayed.