Drop the x value from the raw data series.
It turns out that we don't use the global X value from the raw samples,
except as an optimisation to avoid doing expensive calculations too often,
so there is no need to store it in the raw data series.

This saves about dataLength * 8 bytes * 8 series =  256k of memory and
opens the way to drop the hash container, which could speed up lookups.