calculate basic stats on the filtered data.
The 'basic stats' graph previously used the raw data, instead of the
filtered data, but it was confusing to users and somewhat less useful,
because the raw data usually contains at least one sample with no
packets. The raw data very rarely showed a non-zero minimum.

It is somewhat faster to calculate stats on the shorter filtered data.
The stats array is now modified in place, which the memory allocator
also appreciates.