major js refactor: add JT global to contain jittertrap and modularise it.
The jittertrap javascript code was starting to sprawl all over the
global namespace, with global variables and functions defined in one
file but called from another and so on.

This change implements the Javascript module pattern to place all the JT
code in the JT namespace. Each 'module' has its own space in the JT
object, resulting in better encapsulation (it's clear when one module
uses a variable or function from another module).

A major drawback of this change is that static analysis is crippled and
run-time debugging is also more challenging. Both JSHint and the Chrome
debugger will complain about an undefined global, but not about an
undefined member symbol.