backend: monitor the first 'valid' iface on startup.
This change lets the backend decide which interface to monitor when it
starts, instead of monitoring 'lo' (and assuming that the frontend will
start next and tell it what to do).

It is the first of a sequence of changes with the goal to keep state
in the backend and push state changes to the frontend, so that multiple
frontends can come and go without messing up each other's state.