remove delta bytes/packets series
This change removes the delta bytes/packets series because they were
not as usefull as I'd hoped. They were effectively an average
per-millisecond rate.

This reduces the amount of data kept, the amount of processing done
per websocket message and reduces garbage generation.

Rates are now calculated as needed for the charts and stats
(traps and measurements) and should be more accurate, because it's no
longer an average of averages.