fix histogram calculation and use filtered data.
The histogram was broken in one of the js refactorings, but it also
looks like it could never have worked properly, because the
normalisation calculation (and it's inverse) were wrong.

The histogram now uses the filtered, rate-formatted data, because the
raw data doesn't make sense; it showed an accurate distribution,
but the units didn't match anything else on the page. Now all the units
match.