store filteredData as a rate, because it's always used as a rate.
Previously, commit 9be5c0739bb5ee469330dee89d5b873108872fcd removed
the rate calculation from the "raw data" and defered it to point where
the "filtered data" (the mean for the chartingPeriod) is consumed.

Since the histogram is now using the filtered data with a rate
multiplier, it makes more sense to apply the rate multiplier once and
store all the filtered data in this format (removing the formatting from
the point of consumption, of course).

The "average of averages" problem mentioned in commit 9be5c0739 was
that the "raw data" was stored as a rate (1ms extrapolated to 1s) and
then averaged again to chartPeriod. The change it made was to defer the
rate conversion on the raw data and filtered data until after the mean
has been calculated in updateStats(). It was still calculating the mean
of the filtered data, which is already the mean of the raw data.

With this change, as concerning the "average of averages" problem,
the mean (rate) of the series is still taken as the mean (rate) of the
filtered data, after the filtered data has been converted to a rate.
It's an unquantified (expectedly modest) loss of accuracy tradeoff for
performance.