fix websocket message rate and calc samples per msg from period.
This is part of ongoing work to make the sample period configurable,
to be able to scale to different machine tipes.

Previously the websocket message size was fixed (10 samples per msg)
and the rate would change if the sample rate changed. This patch changes
to a fixed message rate (50Hz) and calculates the number of samples per
message from the sample period and message rate.