calculate mean and std deviation for sample time errors.
This change calculates the mean, max and standard deviation for the
sample time errors (in nanoseconds) per message.

It's meant to give an indication of the quality of the sampling,
so in the tradeoff between higher backend load or higher websocket
and javascript load, it seemed better to do this computation on the
backend.