add a fudge factor to improve measurement accuracy.
The mean delay between the wakeup deadline and the netlink response
from the kernel is about 100us on the mousetrap hardware and around
500us on the T410 thinkpad.

This fudge factor shifts the actual deadline earlier by 100us to
improve the accuracy (but not the precision) of the measurements.
But, since we're not looking at any phase information, it's mostly a
vanity hack to please the code that checks for measurement fidelity.