fix the dev_select message format
The dev_select message is used as a command to the server to change the
selected interface and as a notification to the client that the interface
has changed. Since it carries the same data and has the same name, this
change ensures that it has the same format too.