WIP: compute_thread calculates mean, charts plot something
Each struct jt_msg_stats (and therefore each json message) now contains
the metrics for a single sample in a single time series. The series is
identifiable by the interval attribute.

Eg, for a period of 200ms, there would be several messages:
40 * ival ==   5E6 (ns),
20 * ival ==  10E6,
10 * ival ==  20E6,
 4 * ival ==  50E6,
 2 * ival == 100E6,
 1 * ival == 200E6

This is a work in progress. jittertrap-core.js no longer performs any
filtering calculations, but does not yet store separate time series
and does not yet alocate each message to the correct series.