drop the 5ms and 200ms windows
Step 1 in making JT run on RPi 1 is to do less work.

The 5ms series requires the sampling thread to produce every 5ms,
the compute thread to do calculate min,max,mean and the websocket
thread to do lots of malloc/free to send a json message. Remove that.

The gap length calculation is worse than O(N), so the 200ms series
is expensive. Remove that.