fix bugs with multiple traps of the same type
Previously the trap 'id' was really just a trap type name, so this
changes all the trap 'id' fields to trap 'type' fields.

It also adds a real trap id, to enable multiple traps of the same type.

Lastly, each trap will store the value that triggered it and track the
worst result.