add a trap for the current bitrate
This adds a trap for the current bitrate. That is, the most recently
collected value, where value is the mean for the selected decimation period.

The 'mean', 'max' and 'min' traps are for the mean, max and min of the
charted data, which is currently the last 200 samples of the selected
decimation period.