fix disappearing chart
When netem parameters are set, the server sends a dev_select msg,
which causes the charts to be reset. The chart reset function set the
size of the svg element to 0. This change simply calls the resize function
at the end of the chart reset function. It's a bit of a hack to patch over
the issue of not knowing how big the chart is before it's been drawn.