fix assert when connecting 9th subscriber
The message queue has a maximum number of subscribers, that is set to 8
for the websocket consumers. The 9th subscription failed as expected,
but was not handled correctly in callback_jittertrap(). It caused an assert
when it attempted to use an invalid consumer_id to consume() from the queue.

This fix is to close the connection when subscription failed (return -1;)
and secondly, not to unsubscribe if the consumer_id is invalid.

The downside of this change is that the consumer_id is no longer opaque.
The caller can now check if consumer_id == 0 to know when it's safe to
skip unsubscribe().