replace zeroing macro with null-valued static struct copy
For struct_pkt_record, we initialise all the fields after malloc.
But, initialising the struct in_addr members is not elegant when using the
macro method, so allocate an extra struct pkt_record with static
initialisation and copy the zero-valued struct instead.

It's expected that this method is faster/cleaner than a memset(),
but was not confirmed.