remove redundant hash handles
Each struct flow_hash is only stored in one hash table at a time.
A copy is made for each new flow when it is added to the interval hash.
Therefore, only one hash handle is needed per flow.

Perhaps a better implementation would store the interval counters with each
flow in the main flow table...