keep incomplete & complete flow tables for each interval
This introduces another set of flow interval accounting tables for
completed intervals.

The problem is that counting intervals are not synchronised to each
other or to the read-out interval and incomplete intervals of differing
lengths cannot sensibly be compared to each other.

So by always using the most recently completed interval table, we avoid
calculating rates from almost empty interval tables and other time-sync
issues.