split pcap and terminal io into separate threads.
This is a big fat ugly patch, but as simple as I could make it. It splits
the pcap handling and interval calculations from the ncurses UI into a
separate thread.

This is part of a series of changes to make a self-contained sampling and
interval calc module that can be tested and re-used with a different user
interface, like jittertrap.