fix missing json messages by increasing queue lengths.
Messages were being lost in the mq_stats_produce() callback due to the
queue being full. Using larger maximum queue depths solves that, so the
issue is not with average queue throughput.

It could be something stalling (possibly related to permessage-deflate
that was enabled in commit 1317309376a97b4) or the doubling in messages
per second when toptalk was added. Either way, deeper queues solve it.