split bump the max flow count to 20.
Collecting only N flows per interval is insufficient to plot N flows,
because the short timeframe patterns are different from the long
timeframe patterns we want to see on the chart.

We want to see the top 10 flows on the chart, so collect the top 20.

FIXME: the flow count is defined in multiple places :(
It should have a single build time definition.