move constants from intervals_user.h to CFLAGS
This moves the INTERVAL_COUNT and MAX_FLOW_COUNT constants out of
intervals_user.h and makes them compile-time environment definitions
instead. The Makefile provides the same default values as before, but it
is simpler to override.

The previous mechanism of letting the outer application provide an
overriding intervals_user.h is tricky to do correctly and it's hard to
detect whether it is done correctly, and the resulting bugs are hard to
track down.