improve flow reordering noise
Increase the number of tracked flows, so that the ranking of the top ten
flows is more stable when flows change over time.

When a flow disappears from the tracked flows, its contribution is no
longer tracked, which results in a void in the stacked graph. Since the
stack cannot have any voids, it results in an inaccurate under-reading.